timer: move contents of clock/clock.rs into clock/mod.rs
